pyproject.toml
[tool.poetry]
name = "telegramscrap"
version = "0.1.0"
description = ""
authors = ["msterhuj <gabin.lanore@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Telethon = "^1.25.2"
rich = "^12.6.0"
pymongo = "^4.2.0"
PyYAML = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
